+++ src/lae/allocator.h
@@ -6,8 +6,8 @@
 struct lae_allocator;
 typedef struct lae_allocator lae_allocator;
 
-typedef void * (*lae_allocator_allocate_function)(const int32_t size, const int32_t count);
-typedef void * (*lae_allocator_reallocate_function)(void *, const int32_t size, const int32_t count);
+typedef void * (*lae_allocator_allocate_function)(const size_t size, const size_t count);
+typedef void * (*lae_allocator_reallocate_function)(void *, const size_t size, const size_t count);
 typedef void (*lae_allocator_free_function)(void *);
 
 lae_allocator * lae_allocator_create    (
@@ -18,8 +18,8 @@ void            lae_allocator_release   ( lae_allocator * allocator );
 
 lae_allocator * lae_allocator_default   ();
 
-void *          lae_allocator_allocate      ( const lae_allocator * allocator, const int32_t size, const int32_t count );
-void *          lae_allocator_reallocate    ( const lae_allocator * allocator, void * allocated, const int32_t size, const int32_t count );
+void *          lae_allocator_allocate      ( const lae_allocator * allocator, const size_t size, const size_t count );
+void *          lae_allocator_reallocate    ( const lae_allocator * allocator, void * allocated, const size_t size, const size_t count );
 void            lae_allocator_free          ( const lae_allocator * allocator, void * allocated );
 
 #endif
